Implement richness of information metric
The third metric to be implemented is richness of information as this is
the only metric left which can be implemented without much additional
effort.

The next step will be to rethink the design of the metrics, the data
input etc. When I have figured this out I will probably write more
extensive tests to cover this very functional part of the platform.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>