Rewrite gon data passing
Assigning gon an instance of a custom class leads to making it use the
to_json method for conversion. The disadvantage is that the class
name maps then to a hash of attributes and their values.

For convenience I map the objects to hashes containing only the
attributes and their values without the class name as key.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>