Increase query size
The tire queries are set to a query size of 10 documents by default.
Since I want to compute the metric over all documents I have set the
size to 10.000 which is ridiculous because actually I want to get all
documents and not a fixed number.

I have also added a dataset property to repositories to display the
number of metadata record associated with the repository.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>