Replace Sidekiq with Typhoeus
For the accuracy metric implementation another framework is more fit for
the task. Sidekiq handles a task completely detached. I need to process
the results immediately with the rest of the accuracy metric object.

For executing HTTP requests in parallel another useful framework is
Typhoeus which seems to be based on the libcurl library as well.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>