Implement sentence splitter and adjust tests
I implemented the sentence splitter by using the tokenize method of the
Tactful Tokenizer. The results were unexpected. Hence, it needs to be
investigated whether the single clauses or actual sentences are required
for computing the readability.

I have adjusted the tests to fit the results to which I agree at the
moment anyway.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>