Fix some typos and bugs
Nothing really big, I just had to fix some typos, respectively bugs
which prevented the recently added code from working. I had to remove
the attribute reader as it was preventing the execution of the name
method.

I had to add a split method to the metric name method implementation as
the metric module is processed by the underscore method as well.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>