Process metric name for accessing repository score
There is a general convention clash for the metric names. As a general
name and for URL structuring I use hyphens, as opposed to underscores
which are used for instance variables and in the database.

The clash prevented the JavaScript code from accessing the repository's
metric scores. This commit fixes that.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>