Fix JSON parsing error
The repetitive AJAX request as started by the click event did not finish
successfully as there was a parsing error. The first result includes a
NaN value. The number of total jobs and processed jobs is not processed,
hence a division through 0.0 happens.

I simply added a statement to catch this case. When a job is issued now,
then the progress bar starts to fill without the need to refresh.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>