Create test to check scoring of subfields
I believe that the completeness metrics creates a higher score with
records having more subfields than others. For instance CKAN metadata
record with more resources than another record might reach a far higher
score, which however, does not reflect the nature of this metric.

For that reason I have created another test to cover this behavior. I
also noticed that I assign a completed score for string fields having an
empty string as value. While the metadata harvester I am developing
replaces empty strings with null this is actually a bad behaviour. After
all it destroys evidence of how the fields are filled. It must not be
the task of the metadata harvester to manipulate the content but of the
metric to deal with these kind of values.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>