Add additional case for completeness metric
I had a typo in my unit test which was one reason for the failing unit
test. Another reason was that I did not differ between object, array and
other types before checking the completeness. If the data is an object
or array the counting has to be delegated once more. Otherwise the
completeness can be checked.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>