Make us of idiomatic Maybe monad for Ruby
The gem andand provides an implementation of the Maybe monad in Ruby.
This way I was able to eliminate the helper method used to retrieve the
value of a multiple nested hash.

The previous implementation used exception handling for cases which are
not very exceptional, but rather to be expected. This should be avoided
anyway.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>