Add color range to the bar chart diagram
To give the numbers more meaning I have added a coloring to the bar
chart diagram. The coloring maps from 0 to 100 for the accessibility
results with 0 being more or less the worst and 100 the best. This,
however, can be irritating, since there can be negative values and
values bigger than 100 as well. Still, I think for now it is a good
solution.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>