Start to refactor test and implementation of RoI
I have started to refactor the richness of information metric code and
test. This includes to extract a lot of the code into separate methods
as well as distinguishing between free text and categorical fields.

As part of this I have added a value method which retrieves a value from
a hash based on a dot-notation which I have used in the recently added
accessibility controller, too.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>