Refactor richness of information metric
To complete the refactoring of the richness of information metric for
now I have changed the field indexing. Instead of the previously used
dot notation to describe the fields I am using now arrays instead.

The tf-idf functionality is now covered with three positive examples.
What is still missing, is a more general test which computes the overall
score. In addition to that, the richness of information metric should
also measure the score for categorical values.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>