Create exception to handle the absence of scores
If a metric score was not computed yet for a repository this would
result sooner or later in null pointer exceptions. For that reason
I have created a module to store custom exceptions. A newly created
exception is raised if a repository has no results yet for a given
metric.

Although this seems to violate the principle of: only use exceptions
for exceptional situation. It seems, however, this is how it is done
in Rails.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>