Start to implement record picker for metric views
For better comparison I am starting to implement a record picker. The
idea is to select different metadata records and not only the best and
worst of a certain metric.

With this commit I am starting to implement it on the Richness of
Information metric. Later this will be generalized.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>