Add additional methods for printing tags
In order to avoid some problems with reusing helper methods I have added
another method for printing tags. Obviously, this needs refactoring. But
I ran into problems with the i-tag in HTML. In HTML5 the i-tag cannot be
self-closed which resulted in some awful re-rendering of the source code
which I was not able to explain at first.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>