Use Scroll API to fetch all metadata
The Scroll API is optimized to fetch large amount of documents. For the
metric computation all documents are required which makes the API useful
for this kind of task.

Per shard one request will be executed, hence I am adjusting the size as
well. My ElasticSearch instance is configured with 5 shards. For reasons
unknown diving the total count through 5 does not result in a result set
containing all documents. For that reason, division through 4.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>