Fix test by assigning total datasets value to stub
Since I had not defined a total datasets value for the model stub in the
test, the query of the total method was executed. Without indexes on the
ElasticSearch instance this fails. Hence, I just added this value.

With this change I also added assertions to check the formatting of the
rendered value.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>