Return zero score for unassertable records in RoI
I still get uncaught exeptions in the Richness of Information metric
implementation. This time, it was for a record which had no fields
defined in order to assert a Richness of Information score.

It only seemed reasonable to assign a score of 0.0 here. If there is no
information, then the score can only be the lowest possible.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>