Add status and logger calls to RoI metric
In order to give feedback for the preprocessing of the metadata, which
can be quite time consuming, in the Richness of Information metric, I
have addded a reference to the worker.

This way the status method at can be invoked. In addition I have added
another logger call as well.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>