Fix Accessibility metric diagram
The problem with the diagram was that the data input exceeded the d3.js
configuration. My solution thins out the values. I am not completely
happy with the solution, since it can only be an intermediate solution.
With the current implementation it is still possible that there are 100
values which might be too much.

With all probability I have to think about an alternative way to make
the metric results accessible. In addition this commit also implements
the normalization of the Accessibility results. This commit fixes #1.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>