Refactor code to access inner record if necessary
In order to address issues #8 and #22 I have refactored the views and
their respective controller and helper code. Worker classes needed a
bit rework, too. After all the metric model classes should only deal
with the problem domain and not with the solution domain to which I
would count how to store metric results, too.

The harvester update fixes #8 and this commit fixes #22.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>