Improve Link Checker metric
I have extended the Link Checker metric implementation for two cases.
One is that the location has been moved, namely HTTP 301 and HTTP 302.
In this case I am updating the request and invalidate the previous one.

The other one is the additional URL escaping. This prevents errors when
URLs contain whitespaces or other characters that require escaping. This
commit improved the overall results for all the repositories.

This commit fixes #24.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>