Determine job state to select bar type
In order to differ between the two currently displayed progress bars I
have added more code. For this, I also renamed the state values as set
by the worker.

There is a problem with the repository names and using them to identify
the progress bar via its CSS class as dots are possible inside a class
name but require escaping. To circumvent this problem I now replace dots
with hyphens.

If the worker is in the state compute, the analyze progress bar can be
filled up as the states follow each other sequentially.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>