Create metric report link via helper method
Just delegated the creation of the metric report link into a helper.
Here I am pasting the metric name into the URL. I do not want to use
underscores for the metric identifier if there is a gap, but a hyphen.
Until now I have made that change on the fly. I think it is better to
use a dedicated module method for that task. This way it can be more
easily changed in the future.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>