Refactor search table result linking
I have reimplemented the linking for the result table when doing the
free-text search on the metadata records. Once again I am using the
Bootstrap plugin. I did not work because the stylesheet was not part
of the asset pipeline.

Also I did a lot of JavaScript code refactoring, making the code more
general and modular to realize this feature.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>