Fix parameter setup when choosing all repositories
I am giving the opportunity to call the metric report on all
repositories. For this I had to adjust the code a bit further.
Especially in the repository concern controller I have to check
for this case. For now I am handling this by creating a repository
instance with the name *.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>