Refactor repository average calculation into model
I have moved the method to calculate the repository score into the
repository model class where, to my mind, it belongs anyway. Also
I have renamed it from average_score to just score, because I think
whether the average, the median or some weighting is done should be
left to the implementation.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>