Make JavaScript code for repository map robust
Due to adding more and more often the repositories to the JavaScript
environment by using gon and the repository concern controller I am
facing JavaScript errors more often.

It is better to rely on a distinct namespace through HTML class
attribution. Hence, I have replaced the if check for gon with a
check for the jQuery selector.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>