Upgrade to Rails 4
Due to eager loading of certain files I have decided to prepone the
Rails upgrade. For this I had to change a lot of files, most of them
come from a newly created Rails 4 project. In particular, I had to
change the custom formatter I am using for logging. This closes #32.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>