Complete pie chart to show link checker responses
This commit finishes the implementation of a pie chart visualizing the
distribution of link checker responses. There was a problem if the pie
chart would be rendered on an invisible element. This would lead to a
stacked legend. I have fixed it by creating the pie chart when the tab
is clicked.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>