Remove follow location implementation
Since the library provides already a follow location option I do not
need to implement this myself. I have removed the code and also added
the option to not verify the SSL peer.

In addition, I have changed the (un)escaping of the URL once more as
well as adding checks to prevent exceptions if the worker is undefined.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>