Harden link checker against nitpicky servers
Apparently, a lot of server have a white list for accepted user agents.
The default user agent of Typhoeus is not part of that list. Hence, I am
using the libcurl user agent from now on. Also adding more configuration
parameters, for instance not checking the host anymore when it comes to
SSL certificates.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>