Remove AngularJS again
After evaluating AngularJS I came to the decision to drop it again.
AngularJS makes perfectly sense with Rails, but only if AngularJS is
responsible for the whole front-end and not only a part. For this
project or at least with the deadline of my master's thesis in mind I
will drop this idea for now.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>