Add modal to be used for metric score weighting
I have added a Bootstrap modal which will be used to set the metric
score weighting. Currently, each input gets the placeholder 1.0 which
will be the default value. In the next step I will implement the code
in JavaScript to parse the values to pass them to the metric score
method.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>