Finish implementation of score configuration
With this commit the repository score can now be adjusted by entering
custom weighting for each metric. With this commit I have added more
refactoring to make the score meter class as reusable as possible so
far. For instance I was able to eliminate duplicate code. By declaring
the score as a class attribute it is possible to animate the score meter
from the current position and not from 0.0 all the time. This commit
closes #33.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>