Finish pie chart implementation
I simply had to change the format the data is passed to the D3.js code.
Instead of pure objects a list of objects is required with the keys key
and value. Also I have added a styling to give the pie chart a white
stroke and I have downsized it, too.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>