Add import POST command
I have renamed the import route to importer so I could add a POST route
for the path import. This POST command should be used to trigger the
import of the repositories into the database. For now it will be based
on the default YAML file which is also used in the Metadata Harvester.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>