Change domain model by introducing snapshot entity
It becomes increasingly hard to get the right data when having more than
one snapshot, because it consumes too much time to run over all metadata
of a repository. Since I am interested only in one metadata import at a
time I think it is best to introduce a new abstraction here: snapshots.

A snapshot groups together a list of metadata records associated to the
repository and with a certain date (the harvesting date).

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>