Improve value presentation for completeness view
I have improved the presentation of values. This includes the field
names as well. Instead of humanize() I use now titleize(). Not only
does it removes the underscores, but capitalizes every word.

There are exceptions like URL, so that I had to define a mapping to
get from "url" => "URL". More might follow.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>