Improve header parsing
Until now I used readline to parse the header of a metadata dump.
Unfortunately, some dumps have all the data in one line or only few.
Eventually, this leads to parsing the whole file. The point was to avoid
exactly this. Instead of readline a loop is used in which readchar is
called.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>