Add tokenizer to metric class and filter to worker
Another approach for a word tokenizer was added to the metric class.
For the Aspell library it is important that attached characters like
periods, commas, brackets, etc. are filtered. So far it looks like a
better approach to test spelling is not to check against every word,
but to check against a list of common spelling mistakes.

This commit is also accompanied by a NaN filter of the final score.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>