Rewrite concerns to auto-load necessary objects
With the concerns rewrite I was able to delegate the domain model object
loading completely into the concerns. If a concern is included its model
objects are automatically fetched by trying a list of parameters.

Also had to touch a lot of other files for this. The method for getting
all metrics is now all instead of list, as before. This fits more into
the overall convention.

The routings has been changed one more again, too. The singular resource
of repository maps to the same controller, but it creates all the
default actions which I had to silence with an empty only.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>