Refactor breadcrumb menus
The breadcrumb menus have now a new concept. They are not used anymore
as partials, but as layouts which are filled with dropdown menus. The
dropdown menu details are specified in helper methods. In order to make
this work for metrics, too, I needed to change some of the core class
methods.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>