Implement presentation of language analysis
The language results are now shown in the numbers tab as well. I had to
make many changes for this. I added a gem to retrieve the ISO-639 names
for each language. The table stylesheet rules display the table nicely
even when there is not enough space. Using em for the table cell widths
is a practice I should soon apply to all the other tables as well.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>