Style weighting slider
I have added a grid into the table cells containing the weight slider.
This way I was able to add the weight label above the slider. For now
the minimum value will be 0 and the maximum value will be 10 with steps
set to 1. A higher granularity would not make much sense.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>