Write analyzer for record quality distribution
With this analyzer it is now possible to retrieve the number of records
belonging to which score group. There are ten groups. One for each 10%.

Calculating the metadata records on the fly is too slow, hence the
document score is now updated with every metric computation. This slows
the whole computation though.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>