Add histogram to display quality distribution
A histogram has now been added to the snapshot view. I have commented
out the hiding of the x-axis line. I had hidden it for the bar charts
on the metric view, but I think in the case of the histogram it adds
a lot of value to the design. I might add the line to the metric view's
bar charts again, too.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>