Escape dots from MongoDB field names
A horrible bug appeared when I wanted to save the analysis data from the
license metric. I took hours to find out that MongoDB prohibit dots in
the field names. Of course Mongoid and MongoDB, too, were failing
silently. The dot is now replaced, at least in the license metric, with
a rather ugly unicode version.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>