Replace Rails jQuery gem with vendor files
On my newly updated Chrome 31 I get the annoying message that
event.returnValue is deprecated. For reasons unknown this is
not included in the current Rails jQuery gem. Hence, I have
downloaded my own copy of jQuery.

Actually it is version 2 I have addded. It is a lot smaller
and supports less browser versions.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>