Extend accuracy metric by measuring file size
The file size of a resource is a typical example where the accuracy can
be measured by taking the distance between actual file size and expected
file size.

I have added this type of measure to the accuracy metric making it
hopefully more useful. Both results are now presented in tables.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>