Write first perform metric computation test
Finally, there is also a test which performs a call to Sidekiq to
trigger a whole metric computation. I have added some specifications
covering the score results.

A more significant change is that metrics which have not been calculated
yet will not be set to 0.0 for the average score calculation.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>