Extend CKAN weighting schema and fix minor issues
Just extended the CKAN weighting schema according to the new JSON
schema. Maybe it would be wise to structure the weighting schema by
using each weight as a key to a list of field values. This way the
ranking of importance would be more clearer.

I have also fixed some minor issues. In the analysis manager test
I am now removing the constant to avoid warnings by the Ruby
interpreter.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>