Overhaul eager loading of metric subclasses
By introducing library modules that get included in the metric code I
ran into a loading dependency problem in Rails. Since I was not able
to resolve the loading order in a conflict-free way.

The new solutions drops the inherited method approach and creates the
list of metrics based on the directory content. Although I do not like
this approach I had to iterate the directory's content anyway. So it is
more or less the same solution with less code.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>