Preserve words which would be eliminated otherwise
There is a special case with short texts in the richness of information
metric. In one dataset the description text was a year range. Due to the
splitter regular expression the whole text is eliminated.

This should be revisited at another time. Especially, multiple appliance
of the word tokenizer should not alter the output more than once.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>