Configure cache stores explicitly
There were problems with the cache store. The test environment used data
from the production environment and vica versa. The default caching
system is file store. Since it did not clean it up, somehow the same
cached data was used in the respective environment. Now the cache store
is set explicitly for every environment.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>