Create SVG library file to store SVG styling
SVG styling like filters or gradients need to be reused. The SVG linking
seems only to work in Firefox, Opera and maybe Safari. In Chrome it is
simply not implemented yet.

There is an open ticket since January 2012 so it does not seem that this
will be implemented anytime soon. For now I will dynamically add it to
every page with an AJAX request.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>