Optimize Rails-JavaScript bridge to pass less data
For JavaScript accessing the domain data I have used gon so far and I am
still intending to do that, but I must stop passing mindlessly all data
to gon.

This included a lot of statistical data from the metrics increasing the
average page size to multiple megabytes. So far this was a major reason
for slow page rendering. With this commit I am starting to refactor it
all for better overall performance.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>