Add Jbuilder to tailor entity JSON
With Jbuilder I can further optimize and tailor the JSON output for
certain entities. A Jbuilder JSON template is used to declare what
fields should be used in the result. This way, for instance, I was
able to make the score just another field enabling me to remove the
calculate score function from JavaScript.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>