Write test for new dropdown menu
I hope this time the whole dropdown thing is solved altogether. I have
written a test for the new design. By using two parameters, everything
should be satisfiable.

A entities array with objects and arrays of objects. The objects will
be passed to the link method and the array is the one being iterating
whereas the array object in the entities array is replaced during the
iteration.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>