Move dropdown menu functionality into the caret
In order to turn the breadcrumb navigation into a proper breadcrumb I
have moved the dropdown functionality into the caret. The display link
will be used to navigate directly to the entities.

Repository > Snapshot > Metric

If the user clicks on Snapshot then the page for the selected repository
and snapshot is open. In effect, this means that the user can navigate
vertically through the domain model.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>