Turn metric classes into singletons
I am starting to turn the metric classes into proper active model
entities. For this I will need actual objects as opposed to symbols.
Thus, I chose to use the singleton pattern.

The metrics will be instantiated at the start of the application. The
parameters that have been passed so far to the initializer are now
passed to the configure method.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>