Unify dashboard application and styling code
In order to unify the code as well as styling of the dashboards
displayed on top of the snapshot and metric I have introduced a
bit of new code.

The loading of external data is now encapsulated. The function
provides the displaying and destroying of the loading spinner.

I have copied a bit of view code to the metric view and created
a new stylesheet to organize overarching dashboard properties.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>