Improve statistic tab of the completeness metric
The raw presentation of the completed fields has been replaced with a
proper table. For this I had to pre-process the data. In fact, I have
added a method to flatten the hash and join the keys afterwards.

In addition, I have added Font Awesome for more icons I can use to style
the dashboard headings.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>