Start to implement new record selection concept
The new metadata record selection should happen as follows. In the
dropdown menu the score ranges are listed as entries, from 0 to 100
in steps of 10.

If an entry is selected the modal appears and lists the records to
select them. This way it is easier to select records from different
score ranges.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>