Finish new record selection implementation
So far it is done and works properly. There were tons of bugs and
changes I had to tackle so I do not feel very confident that this
will work stable in the future, but this is true for a lot of parts
in the code base. Refactoring is a required consequence in any case.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>