Fix results of quality distribution for metrics
Due to the refactoring of the metric object the quality distribution was
broken. As always the simple fix was by fetching the actual identifier
and adding it to the rest of the code.

I have to keep in mind what will happen when I change the metric model
again. For instance, to one where the metric objects are metric classes.
It might not be a problem, because in the end there will be an id
method, too.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>