Fix metric view for weighted completeness
The weighted completeness view can now be accessed again without errors,
too. On the one hand I had to fix the code for looking up the partial
and the analyzer and on the other hand I forgot to add another isPath
directive in the JavaScript so the Treemap is not only instantiated on
the completeness view, but also the weighted completeness view.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>